finalize renaming
