+++ DESCRIPTION
@@ -76,8 +76,8 @@ Collate:
     'parquet.R'
     'pkg-knitr.R'
     'pkg-nanoarrow.R'
-    'rerr.R'
     'rlang.R'
+    'rpolarserr.R'
     'rust_result.R'
     's3_methods.R'
     'series__series.R'
