+++ man/dot-DollarNames.RPolarsErr.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/rerr.R
+% Please edit documentation in R/rpolarserr.R
 \name{.DollarNames.RPolarsErr}
 \alias{.DollarNames.RPolarsErr}
 \title{auto complete $-access into a polars object}
