+++ man/result.Rd
@@ -29,12 +29,12 @@ Capture any R error and return a rust-like Result
  throw_simpleError  = \() stop("Imma simple error")
  result(throw_simpleError())
 
- throw_Rerr = \() unwrap(
+ throw_RPolarsErr = \() unwrap(
   Err(.pr$RPolarsErr$new()$bad_robj(42)$mistyped("String")$when("doing something"))
  )
- res_Rerr = result(throw_Rerr())
- str(res_Rerr)
- RPolarsErr = unwrap_err(res_Rerr)
+ res_RPolarsErr = result(throw_RPolarsErr())
+ str(res_RPolarsErr)
+ RPolarsErr = unwrap_err(res_RPolarsErr)
  RPolarsErr$contexts()
 }
 \keyword{internal}
